Time: 95 ms (62.02%), Space: 77 MB (22.4%) - LeetHub

+++ 1894-find-the-student-that-will-replace-the-chalk/1894-find-the-student-that-will-replace-the-chalk.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int chalkReplacer(vector<int>& chalk, int k) {
-        int useInOneItration = 0;
+        long long useInOneItration = 0;
         for(auto & x : chalk)
             useInOneItration += x;
         if(k>=useInOneItration)
